+++ includes/services/class-recipe-access-service.php
@@ -5,7 +5,7 @@
  * @package MealPlannr
  */
 
-namespace MealPlannr;
+namespace MealPlannr\Services;
 
 /**
  * Recipe Access Service
